api done

+++ app/src/main/java/com/example/cse438/cse438_assignment2/Data/TrendingResult.kt
@@ -15,6 +15,6 @@ data class TrendingResult(
     val adult: Boolean,
     val overview: String,
     val poster_path: String,
-    val popularity: Long,
+    val popularity: Double,
     val media_type: String
 )
